Integrated nightwatch end2end testing into travis along with test example
