+++ .travis.yml
@@ -1,9 +1,13 @@
 language: node_js
 node_js:
-- '6'
+  - '6'
 before_script:
-- npm install -g gulp
+  - npm install -g gulp
+  - python -m SimpleHTTPServer &
+  - sleep 2
 script: gulp
+addons: 
+  sauce_connect: true
 deploy:
   skip-cleanup: true
   provider: heroku
