+++ reports/CHROME_58.0.3029.81_LINUX_example.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<testsuites errors="0"
+            failures="0"
+            tests="1">
+
+  <testsuite name="example"
+    errors="0" failures="0" hostname="" id="" package="example" skipped="0"
+    tests="1" time="4.606" timestamp="Mon, 01 May 2017 18:05:21 GMT">
+  
+    <testcase name="Guinea Pig Assert Title" classname="example" time="4.606" assertions="2">
+
+    
+
+    
+    </testcase>
+  
+  
+  </testsuite>
+</testsuites>
