+++ test/nightwatch/hw2_e2e.js
@@ -1,0 +1,16 @@
+var config = require('../../nightwatch.conf.BASIC.js');
+
+module.exports = {
+  'Super Time Validator - Startup Test': function(browser) {
+    browser
+      .url('http://maoproj.herokuapp.com/')
+      .waitForElementVisible('body')
+      .assert.title('Super Time Validator')
+      .assert.visible('#inputTime')
+      .assert.visible('#validateButton')
+      .assert.visible('.switch')
+      .assert.hidden('#resultDisplay')
+      .saveScreenshot('screenshots/super_time_validator.png')
+      .end();
+  }
+};
